feat: extract into a component
